Added dask and numpy percentile implementation.
This replicates the behaviour of datacube_statistician argpercentile which uses index rounding instead of the more standard linear interpolation.

A caveat of this implementation is that it assumes that nodata is either the highest value, lowest value, or NaN.

np.nanpercentile was not used for performance reasons - this implementation is ~700x faster.