Working geomedian_with_mads!!
Because this Rust implementation computes both geomedian
and MADs before returning to Python. The geomedian result
can be either floats or ints, and the MADs results are always
floats. Since you can't store both floats and ints in the same
array it returns two arrays.

This required a pretty major rejiggering of the geomedian_with_mads
function. Moving most of the logic for getting the data out of the
xr.Dataset in the correct shape and axis ordering, and then putting it
back into an xr.Dataset on completion.

I think this leaves a whole chunk of code which can be deleted, but
need to do more testing first.